Enable Functions SDK debug mode. (#3873)
Functions SDK now includes a debug mode that can be enabled via environment variables to operate in ways that's useful during development. See https://github.com/firebase/firebase-functions/pull/983 for an example.

We will turn on Debug Mode and enable `skipTokenVerification` feature on the Functions Emulator. This means that callable functions will decode but not verify the signature of Auth and App Check tokens which is more or less what the monkey-patching of Firebase Functions SDK achieves. For some function triggers, we will now rely on the debug mode instead of the monkey patching to bypass the token verification.